requirements.txt
gradio
requests
langchain
python-dotenv
langchain-openai